refactor: eliminate repetition in quiz category components - move all quiz data (questions, answers & options) into a new singular file (CategoriesData.jsx) - render a single component in Categories.jsx as opposed to 7 distinct components for each category.
