--- src/components/Resources/Quiz/Categories/CBQ/CBQData.jsx
@@ -1,76 +1,0 @@
-export const CBQQuestions = [
-    {
-        questionText: "Which of the following should you do to restrict access to your files and devices?",
-        answerOptions: [
-            { answerText: "Update your software once a year.", isCorrect: false },
-            {
-                answerText: "Share passwords only with colleagues you trust.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Have your staff members access information via an open Wi-Fi network.",
-                isCorrect: false,
-            },
-            { answerText: "Use multi-factor authentication.", isCorrect: true },
-        ],
-    },
-    {
-        questionText:
-            "Backing up important files offline, on an external hard drive or in the cloud, will help protect your business in the event of a cyber attack. True or False?",
-        answerOptions: [
-            { answerText: "True", isCorrect: true },
-            { answerText: "False", isCorrect: false },
-        ],
-    },
-    {
-        questionText:
-            "Which is the best answer for which people in a business should be responsible for cybersecurity?",
-        answerOptions: [
-            {
-                answerText:
-                    "Business owners. They run the business, so they need to know cybersecurity basics and put them in practice to reduce the risk of cyber attacks.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "IT specialists, because they are in the best position to know about and promote cybersecurity within a business.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "Managers, because they are responsible for making sure that staff members are following the right practices.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "All staff members should know some cybersecurity basics to reduce the risk of cyber attacks.",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText: "Cyber criminals only target large companies?",
-        answerOptions: [
-            { answerText: "True", isCorrect: false },
-            { answerText: "False", isCorrect: true },
-        ],
-    },
-    {
-        questionText: "Which of the following is the best answer for how to secure your router?",
-        answerOptions: [
-            {
-                answerText: "Change the default name and password of the router.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Turn off the router’s remote management.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Log out as the administrator once the router is set up.",
-                isCorrect: false,
-            },
-            { answerText: "All of the above.", isCorrect: true },
-        ],
-    },
-];
