--- src/components/Resources/Quiz/Categories/PSQ/PSQData.jsx
@@ -1,83 +1,0 @@
-export const PSQQuestions = [
-    {
-        questionText: "Promoting physical security includes protecting:",
-        answerOptions: [
-            {
-                answerText: "Only paper files.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Only paper files and any computer on which you store electronic copies of those files.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Only paper files, flash drives, and point-of-sale devices.",
-                isCorrect: false,
-            },
-            {
-                answerText: "All the above plus any other device with sensitive information on it.",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Paper files that have sensitive information should be disposed of in a locked trash bin. True or False?",
-        answerOptions: [
-            { answerText: "True", isCorrect: false },
-            { answerText: "False", isCorrect: true },
-        ],
-    },
-    {
-        questionText:
-            "When you hit the “delete” key, that means a file is automatically removed from your computer. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: false,
-            },
-            {
-                answerText: "False",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText: "Which one of these statements is true?",
-        answerOptions: [
-            {
-                answerText:
-                    " It's best to use multi-factor authentication to access areas of the business network with sensitive information.",
-                isCorrect: true,
-            },
-            {
-                answerText:
-                    "You should use the same password for key business devices to guarantee that high-level employees can access them in an emergency.",
-                isCorrect: false,
-            },
-            {
-                answerText: "The best way to protect business data is to make sure no one loses any device.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "You shouldn't limit login attempts on key business devices, because getting locked out for having too many incorrect attempts would leave you unable to access your accounts.",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Only people with access to sensitive data need to be trained on the importance of the physical security of files and equipment. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: false,
-            },
-            {
-                answerText: "False",
-                isCorrect: true,
-            },
-        ],
-    },
-];
