--- src/components/Resources/Quiz/Categories/Phishing/Phishing.jsx
@@ -1,60 +1,0 @@
-import React from "react";
-import { PhishingQuestions } from "./PhishingData";
-import RenderProgressIndicator from "../../../../../utils/components/RenderProgressIndicator";
-
-export default function Phishing({
-    AnswerSection,
-    QuestionButton,
-    QuestionCount,
-    QuestionSection,
-    QuestionText,
-    QuizBody,
-    QuizSection,
-    ResetButton,
-    ScoreInfo,
-    ScoreSection,
-    currentQuestion,
-    showScore,
-    score,
-    // scoreList,
-    handleAnswerButtonClick,
-    handleResetButton,
-}) {
-    return (
-        <QuizSection>
-            {showScore ? (
-                <ScoreSection>
-                    <ScoreInfo>
-                        You scored {score} out of {PhishingQuestions.length}
-                    </ScoreInfo>
-                    <ResetButton onClick={() => handleResetButton(score)}>Start again</ResetButton>
-                </ScoreSection>
-            ) : (
-                <QuizBody>
-                    <QuestionSection>
-                        <QuestionCount>
-                            <RenderProgressIndicator
-                                questionsArray={PhishingQuestions}
-                                currentQuestion={currentQuestion}
-                            />
-                            <span>Question {currentQuestion + 1}</span>
-                        </QuestionCount>
-                        <QuestionText>{PhishingQuestions[currentQuestion].questionText}</QuestionText>
-                    </QuestionSection>
-                    <AnswerSection>
-                        {PhishingQuestions[currentQuestion].answerOptions.map((answerOption, i) => (
-                            <QuestionButton
-                                key={i}
-                                onClick={() =>
-                                    handleAnswerButtonClick(answerOption.isCorrect, PhishingQuestions.length)
-                                }
-                            >
-                                {answerOption.answerText}
-                            </QuestionButton>
-                        ))}
-                    </AnswerSection>
-                </QuizBody>
-            )}
-        </QuizSection>
-    );
-}
