--- src/components/Resources/Quiz/Categories/Ransomware/RansomewareData.jsx
@@ -1,82 +1,0 @@
-export const RansomwareQuestions = [
-    {
-        questionText: "What is ransomware?",
-        answerOptions: [
-            {
-                answerText:
-                    "Software that infects computer networks and mobile devices to hold your data hostage until you send the attackers money.",
-                isCorrect: true,
-            },
-            {
-                answerText: "Computer equipment that criminals steal from you and won't return until you pay them.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Software used to protect your computer or mobile device from harmful viruses.",
-                isCorrect: false,
-            },
-            {
-                answerText: "A form of cryptocurrency.",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Local backup files saved on your computer will protect your data from being lost in a ransomware attack. True or False?",
-        answerOptions: [
-            { answerText: "True", isCorrect: false },
-            { answerText: "False", isCorrect: true },
-        ],
-    },
-    {
-        questionText: "Which of these best describes how criminals start ransomware attacks?",
-        answerOptions: [
-            {
-                answerText: "Sending a scam email with links or attachments that put your data and network at risk.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Getting into your server through vulnerabilities and installing malware.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "Using infected websites that automatically download malicious software to your computer or mobile device.",
-                isCorrect: false,
-            },
-            {
-                answerText: "All of the above.",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText:
-            "If you encounter a ransomware attack, the first thing you should do is pay the ransom. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: false,
-            },
-            {
-                answerText: "False",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Setting your software to auto-update is one way you can help protect your business from ransomware. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: true,
-            },
-            {
-                answerText: "False",
-                isCorrect: false,
-            },
-        ],
-    },
-];
