--- src/components/Resources/Quiz/Categories/SRAQ/SRAQData.jsx
@@ -1,88 +1,0 @@
-export const SRAQQuestions = [
-    {
-        questionText:
-            "Before connecting remotely to the company network, your personal device should meet the same security requirements as company-issued devices. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: true,
-            },
-            {
-                answerText: "False",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText: "What is a common way to help protect devices connected to the company network?",
-        answerOptions: [
-            {
-                answerText: "Only use laptops and other mobile devices with full-disk encryption.",
-                isCorrect: true,
-            },
-            {
-                answerText:
-                    "Change your smartphone settings to let your devices connect automatically to public Wi-Fi.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Let guests and customers use the same secure Wi-Fi that you use.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Use the router's pre-set password so you won't forget it.",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Keeping your router's default name will help security professionals identify it and thus help protect your network's security. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: false,
-            },
-            {
-                answerText: "False",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText:
-            "WPA2 and WPA3 encryption are the encryption standards that will protect information sent over a wireless network. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: true,
-            },
-            {
-                answerText: "False",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Which of the following describes the best way to make sure you are securely accessing the company network remotely?",
-        answerOptions: [
-            {
-                answerText: "Read your company's cybersecurity policies thoroughly.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Use VPN when connecting remotely to the company network.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Use unique, complex network passwords and avoid unattended, open workstations.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Do all of the above.",
-                isCorrect: true,
-            },
-        ],
-    },
-];
