--- src/components/Resources/Quiz/Categories/TSSQ/TSSQData.jsx
@@ -1,93 +1,0 @@
-export const TSSQQuestions = [
-    {
-        questionText: "Which of the following scenarios does NOT describe a tech support scam?",
-        answerOptions: [
-            {
-                answerText:
-                    "Someone calls and tells you they've found viruses on your computer, then asks for credit card information so they can bill you for tech support services.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "While you're browsing online, an urgent message pops up telling you that there's a problem with your computer and directs you to a website to pay for tech support services.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "A caller asks you to give him remote access to your computer to fix a problem in your computer.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "You pay a trusted security professional to check your network for intrusions, and the professional tells you that your network has a problem that needs to be fixed.",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText:
-            "True or False? You can avoid scams by only taking tech support calls from well-known tech companies.",
-        answerOptions: [
-            { answerText: "True", isCorrect: false },
-            { answerText: "False", isCorrect: true },
-        ],
-    },
-    {
-        questionText: "Which of these answers describes the best way to protect against tech support scams?",
-        answerOptions: [
-            {
-                answerText: "Use a unique password for each account.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Scan your computer for any unknown software.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Hang up on callers who say your computer has a problem.",
-                isCorrect: false,
-            },
-            {
-                answerText: "All of the above.",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText:
-            "True or False? Small businesses should focus more on other cybersecurity threats, because tech support scammers usually target only large companies.",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: false,
-            },
-            {
-                answerText: "False",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText: "Which is the best way to protect against viruses or other security threats?",
-        answerOptions: [
-            {
-                answerText:
-                    "Call your security software company to review the steps it has taken to set up virus protection and what else it has done or can do to protect your business.",
-                isCorrect: true,
-            },
-            {
-                answerText:
-                    "Hire a new company that has made the effort to alert you to viruses on your system and offers to help you fix them.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Install new virus protection software that you find online.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Change the network password.",
-                isCorrect: false,
-            },
-        ],
-    },
-];
