--- src/components/Resources/Quiz/Categories/VSQ/VSQData.jsx
@@ -1,102 +1,0 @@
-export const VSQQuestions = [
-    {
-        questionText:
-            "What steps should you take when selecting vendors who will have access to your sensitive information? Pick the best answer.",
-        answerOptions: [
-            {
-                answerText:
-                    "Include provisions for security in your vendor contracts, like a plan to evaluate and update security controls.",
-                isCorrect: true,
-            },
-            {
-                answerText: "Only do business with well-known vendors.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Ensure that your vendors understand your compliance rules.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Confirm that the vendor understands the importance of cybersecurity.",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Anyone with access to your business network should be required to use a strong password. How long should a strong password be?",
-        answerOptions: [
-            {
-                answerText:
-                    "Passwords should be at least 8 characters with a mix of numbers, symbols, and both capital and lowercase letters.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "Passwords should be at least 5 characters with a mix of numbers, symbols, and both capital and lowercase letters.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "Passwords should be at least 12 characters with a mix of numbers, symbols, and both capital and lowercase letters.",
-                isCorrect: true,
-            },
-            {
-                answerText:
-                    "Passwords should be at least 10 characters with a mix of numbers, symbols, and both capital and lowercase letters.",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Requiring vendors to use multi-factor authentication to access your network makes users take an additional step beyond logging in with a password. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: true,
-            },
-            {
-                answerText: "False",
-                isCorrect: false,
-            },
-        ],
-    },
-    {
-        questionText:
-            "Properly configured strong encryption, recommended for any devices that connect remotely to your network, can help you detect cyber attacks in your system. True or False?",
-        answerOptions: [
-            {
-                answerText: "True",
-                isCorrect: false,
-            },
-            {
-                answerText: "False",
-                isCorrect: true,
-            },
-        ],
-    },
-    {
-        questionText:
-            "What should you do if a vendor has a breach that impacts your business data? Pick the best answer.",
-        answerOptions: [
-            {
-                answerText: "Change all network passwords.",
-                isCorrect: false,
-            },
-            {
-                answerText: "Turn off all your computers and devices.",
-                isCorrect: false,
-            },
-            {
-                answerText:
-                    "Make sure that the vendor fixes the vulnerabilities and ensures that your information will be safe going forward.",
-                isCorrect: true,
-            },
-            {
-                answerText: "Disable multi-factor authentication systems.",
-                isCorrect: false,
-            },
-        ],
-    },
-];
